refactor : PieceType의 패키지 변경
