+++ src/test/java/chess/repository/FakeBoardDao.java
@@ -2,7 +2,7 @@ package chess.repository;
 
 import chess.domain.piece.Piece;
 import chess.domain.piece.PieceColor;
-import chess.domain.piece.PieceType;
+import chess.dto.PieceType;
 import chess.domain.piece.type.*;
 import chess.domain.position.Position;
